pyproject.toml
[project]
name = "valthos-catalog"
version = "0.1.0"
description = "Genomic Catalog POC using LanceDB"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "lancedb>=0.4.0",
    "biopython>=1.80",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
